refactor: Add CourseCard component
This commit adds a new component called CourseCard. The CourseCard component is responsible for rendering a card that displays information about a course, including the course name, instructor, day, time, and progress. It also includes a button to view more details about the course. This component makes use of the Badge, Button, Card, CardContent, CardFooter, and CardTitle components from the ui module, as well as icons from the lucide-react library.