feat: Add useSession hook for fetching session data
This commit adds a new file, use-session.ts, which contains the useSession hook. This hook utilizes the SWR library to fetch session data from the "/api/sessions" endpoint. The fetched data is then returned by the hook.

The useSession hook will be used to retrieve session information for authenticated users in the application.

The commit also includes a minor update to api.ts, adding a fetcher function that parses the response as JSON.

Lastly, a new file, api.ts, is added to the repository. This file defines the SessionData type, which represents the structure of session data returned by the API.